Square pixel size on Scala's side is now inferred.
The addon draws three squares at the top left of the screen: one black, one white and one black. The algorithm which decodes the game now look for the first and white pixel that it sees, and infers the size (in pixel) of the squares.

This was needed because the actual size of the squares are sometimes different than the size the addon uses.